
\name{hs-chrs}
  
\alias{hs-chrs}
\alias{hsToplevel}
\alias{hsAutosomes}
\alias{hsAllosomes}
\alias{hsLinear}

\title{Human Chromosome Names}

\description{

  List human chromosome names.

}
  
\usage{
hsToplevel()
hsAutosomes()
hsAllosomes()
hsLinear()
}

\value{

  Character vector with chromosome names (NCBI notation).

}
  
\examples{
hsToplevel()

hsAutosomes()

hsAllosomes()

hsLinear()
}
