\name{MutationalSignatures}

\docType{class}
      
\alias{MutationalSignatures}
\alias{MutationalSignatures-class}
\alias{signatures}
\alias{signatures,MutationalSignatures-method}
\alias{samples}
\alias{samples,MutationalSignatures-method}
\alias{observed}
\alias{observed,MutationalSignatures-method}
\alias{fitted}
\alias{fitted,MutationalSignatures-method}
\alias{show}
\alias{show,MutationalSignatures-method}

\title{'MutationalSignatures' class and methods}

\description{

  Object representing of somatic signatures.
    
}

\usage{

\S4method{signatures}{MutationalSignatures}(object)

\S4method{samples}{MutationalSignatures}(object)

\S4method{observed}{MutationalSignatures}(object)

\S4method{fitted}{MutationalSignatures}(object)

\S4method{show}{MutationalSignatures}(object)

}

\arguments{

  \item{object}{'MutationalSignatures' object}

}

\seealso{

  \code{\link{identifySignatures}}
 
}

\value{

  help("MutationalSignatures")

}
