\name{sca-data}

\alias{sca-data}
\alias{sca_sigs}
\alias{sca_mm}
\alias{sca_motifs_tiny}
\alias{sigs_nmf}
\alias{sigs_pca}

\docType{data}

\title{SomaticCancerAlterations Results}

\description{

  Motif matrix and 5 estimated signatures (NMF) from the somatic variant
  calls in the 'SomaticCancerAlterations' package.  For details, see the
  vignette of the 'SomaticSignatures' package.
  
}

\seealso{

  SomaticCancerAlterations package

}

\examples{

data(sca_motifs_tiny, package = "SomaticSignatures")

data(sca_mm, package = "SomaticSignatures")

data(sca_sigs, package = "SomaticSignatures")  

}

\keyword{datasets}
