\name{signatures21-data}

\alias{signatures21-data}
\alias{signatures21}

\docType{data}

\title{21 Signatures}

\description{

  Published signatures, taken from \url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/signatures.txt}
  
}

\references{

  Alexandrov, Ludmil B., Serena Nik-Zainal, David C. Wedge, Samuel A. J. R. Aparicio, Sam Behjati, Andrew V. Biankin, Graham R. Bignell, et al. Signatures of Mutational Processes in Human Cancer. Nature 500, no. 7463 (August 22, 2013): 415-21. doi:10.1038/nature12477.

}
  
\examples{

data(signatures21, package = "SomaticSignatures")  

head(signatures21)  

}

\keyword{datasets}
