\name{kmers-data}

\alias{kmers-data}
\alias{kmers}
\alias{k3wg}
\alias{k3we}

\docType{data}

\title{Kmer datasets}

\description{

  3mer base frequencies of human whole-genome and whole-exome sampling,
  based on the hg19/GRCh37 reference sequence.

  For details, see the 'inst/scripts/kmers-data.R' script.
  
}
  
\seealso{

  \code{\link{kmerFrequency}}
  
}

\value{

  Vectors with frequency of k-mers.

}

\examples{

data(kmers, package = "SomaticSignatures")  

}

\keyword{datasets}
