\name{sca-data}

\alias{sca-data}
\alias{sca_sigs}
\alias{sca_mm}
\alias{sigs_nmf}
\alias{sigs_pca}

\docType{data}

\title{SomaticCancerAlterations Results}

\description{

  Motif matrix and 5 estimated signatures (NMF) from the
  'SomaticCancerAlterations' dataset.  For details, see the vignette of
  the 'SomaticSignatures' package.
  
}

\seealso{

  'SomaticCancerAlterations' package

}

\examples{

data(sca_mm, package = "SomaticSignatures")

data(sca_sigs, package = "SomaticSignatures")  

}

\keyword{datasets}
